dom/base/test/test_websocket1.html

<!DOCTYPE HTML>
<html>
<head>
  <meta http-equiv="Content-Type" content="text/html;charset=UTF-8"></meta>
  <title>WebSocket test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="websocket_helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="testWebSocket()">
<script class="testbody" type="text/javascript">

function test1() {
  return new Promise(function(resolve, reject) {
    try {
      var ws = CreateTestWS("http://mochi.test:8888/tests/dom/base/test/file_websocket");
      ok(false, "test1 failed");
    } catch (e) {
      ok(true, "test1 failed");
    }

    resolve();
  });
}

// this test expects that the serialization list to connect to the proxy
// is empty.
function test2() {
  return new Promise(function(resolve, reject) {
    var waitTest2Part1 = true;
    var waitTest2Part2 = true;

    var ws1 = CreateTestWS("ws://sub2.test2.example.com/tests/dom/base/test/file_websocket", "test-2.1");
    var ws2 = CreateTestWS("ws://sub2.test2.example.com/tests/dom/base/test/file_websocket", "test-2.2");

    var ws2CanConnect = false;

    function maybeFinished() {
      if (!waitTest2Part1 && !waitTest2Part2) {
        resolve();
      }
    }

    ws1.onopen = function() {
      ok(true, "ws1 open in test 2");
      ws2CanConnect = true;
      ws1.close();
    }

    ws1.onclose = function(e) {
      waitTest2Part1 = false;
      maybeFinished();
    }

    ws2.onopen = function() {
      ok(ws2CanConnect, "shouldn't connect yet in test-2!");
      ws2.close();
    }

    ws2.onclose = function(e) {
      waitTest2Part2 = false;
      maybeFinished();
    }
  });
}

function test3() {
  return new Promise(function(resolve, reject) {
    var hasError = false;
    var ws = CreateTestWS("ws://this.websocket.server.probably.does.not.exist");

    ws.onopen = shouldNotOpen;

    ws.onerror = function (e) {
      hasError = true;
    }

    ws.onclose = function(e) {
      shouldCloseNotCleanly(e);
      ok(hasError, "rcvd onerror event");
      is(e.code, 1006, "test-3 close code should be 1006 but is:" + e.code);
      resolve();
    }
  });
}

function test4() {
  return new Promise(function(resolve, reject) {
    try {
      var ws = CreateTestWS("file_websocket");
      ok(false, "test-4 failed");
    } catch (e) {
      ok(true, "test-4 failed");
    }

    resolve();
  });
}

function test5() {
  return new Promise(function(resolve, reject) {
    try {
      var ws = CreateTestWS("ws://mochi.test:8888/tests/dom/base/test/file_websocket", "");
      ok(false, "couldn't accept an empty string in the protocol parameter");
    } catch (e) {
      ok(true, "couldn't accept an empty string in the protocol parameter");
    }

    try {
      var ws = CreateTestWS("ws://mochi.test:8888/tests/dom/base/test/file_websocket", "\n");
      ok(false, "couldn't accept any not printable ASCII character in the protocol parameter");
    } catch (e) {
      ok(true, "couldn't accept any not printable ASCII character in the protocol parameter");
    }

    try {
      var ws = CreateTestWS("ws://mochi.test:8888/tests/dom/base/test/file_websocket", "test 5");
      ok(false, "U+0020 not acceptable in protocol parameter");
    } catch (e) {
      ok(true, "U+0020 not acceptable in protocol parameter");
    }

    resolve();
  });
}

function test6() {
  return new Promise(function(resolve, reject) {
    var ws = CreateTestWS("ws://mochi.test:8888/tests/dom/base/test/file_websocket", "test-6");
    var counter = 1;

    ws.onopen = function() {
      ws.send(counter);
    }

    ws.onmessage = function(e) {
      if (counter == 5) {
        is(e.data, "あいうえお", "test-6 counter 5 data ok");
        ws.close();
      } else {
        is(parseInt(e.data), counter+1, "bad counter");
        counter += 2;
        ws.send(counter);
      }
    }

    ws.onclose = function(e) {
      shouldCloseCleanly(e);
      resolve();
    }
  });
}

function test7() {
  return new Promise(function(resolve, reject) {
    var ws = CreateTestWS("ws://sub2.test2.example.org/tests/dom/base/test/file_websocket", "test-7");
    var gotmsg = false;

    ws.onopen = function() {
      ok(true, "test 7 open");
    }

    ws.onmessage = function(e) {
      ok(true, "test 7 message");
      is(e.origin, "ws://sub2.test2.example.org", "onmessage origin set to ws:// host");
      gotmsg = true;
      ws.close();
    }

    ws.onclose = function(e) {
      ok(gotmsg, "recvd message in test 7 before close");
      shouldCloseCleanly(e);
      resolve();
    }
  });
}

function test8() {
  return new Promise(function(resolve, reject) {
    var ws = CreateTestWS("ws://mochi.test:8888/tests/dom/base/test/file_websocket", "test-8");

    ws.onopen = function() {
      is(ws.protocol, "test-8", "test-8 subprotocol selection");
      ws.close();
    }

    ws.onclose = function(e) {
      shouldCloseCleanly(e);
      // We called close() with no close code: so pywebsocket will also send no
      // close code, which translates to code 1005
      is(e.code, 1005, "test-8 close code has wrong value:" + e.code);
      is(e.reason, "", "test-8 close reason has wrong value:" + e.reason);
      resolve();
    }
  });
}

function test9() {
  return new Promise(function(resolve, reject) {
    var ws = CreateTestWS("ws://test2.example.org/tests/dom/base/test/file_websocket", "test-9");

    ws._receivedErrorEvent = false;

    ws.onopen = shouldNotOpen;

    ws.onerror = function(e) {
      ws._receivedErrorEvent = true;
    }

    ws.onclose = function(e) {
      ok(ws._receivedErrorEvent, "Didn't received the error event in test 9.");
      shouldCloseNotCleanly(e);
      resolve();
    }

    ws.close();
  });
}

function test10() {
  return new Promise(function(resolve, reject) {
    var ws = CreateTestWS("ws://sub1.test1.example.com/tests/dom/base/test/file_websocket", "test-10");

    ws.onclose = function(e) {
      shouldCloseCleanly(e);
      resolve();
    }

    try {
      ws.send("client data");
      ok(false, "Couldn't send data before connecting!");
    } catch (e) {
      ok(true, "Couldn't send data before connecting!");
    }

    ws.onopen = function()
    {
      ok(true, "test 10 opened");
      ws.close();
    }
  });
}

var tests = [
  test1,  // client tries to connect to a http scheme location;
  test2,  // assure serialization of the connections;
  test3,  // client tries to connect to an non-existent ws server;
  test4,  // client tries to connect using a relative url;
  test5,  // client uses an invalid protocol value;
  test6,  // counter and encoding check;
  test7,  // onmessage event origin property check
  test8,  // client calls close() and the server sends the close frame (with no
          // code or reason) in acknowledgement;
  test9,  // client closes the connection before the ws connection is established;
  test10, // client sends a message before the ws connection is established;
];

function testWebSocket() {
  doTest();
}

</script>

<div id="feedback">
</div>

</body>
</html>
